requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
joblib
jupyter
 xgboost
imbalanced-learn
fastapi
uvicorn[standard]
pydantic
pyyaml
shap
